refactor: split continuum artifact json adapter
